requirements.txt
flask==3.0.0
azure-cognitiveservices-speech==1.40.0
openai==1.54.4
python-dotenv==1.0.0
httpx==0.27.0
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
faiss-cpu==1.9.0
tiktoken==0.8.0